+++ src/BlockTracker.ts
@@ -1,0 +1,11 @@
+import type SafeEventEmitter from '@metamask/safe-event-emitter';
+
+export type BlockTracker = SafeEventEmitter & {
+  destroy(): Promise<void>;
+
+  isRunning(): boolean;
+
+  getCurrentBlock(): string | null;
+
+  getLatestBlock(): Promise<string>;
+};
