+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './PollingBlockTracker';
 export * from './SubscribeBlockTracker';
+export * from './BlockTracker';
